refactor: move 2fa form
